conditionally request assets

+++ packages/dos-donts-block/src/DosDontsBlock.tsx
@@ -6,7 +6,6 @@ import {
     rgbStringToRgbObject,
     useAssetChooser,
     useAssetUpload,
-    useBlockAssets,
     useBlockSettings,
     useEditorState,
     useFileInput,
@@ -22,7 +21,7 @@ import {
     joinClassNames,
     useDndSensors,
 } from '@frontify/guideline-blocks-settings';
-import { FC, useEffect, useRef, useState } from 'react';
+import { FC, useContext, useEffect, useRef, useState } from 'react';
 import { DoDontItem, SortableDoDontItem } from './DoDontItem';
 import { BlockMode, ChangeType, DoDontType, GUTTER_VALUES, Item, Settings, ValueType } from './types';
 import {
@@ -37,13 +36,31 @@ import {
     PatternTheme,
     generateRandomId,
 } from '@frontify/fondue';
+import { AssetsContext, AssetsProvider, NoAssetsProvider } from './AssetsProvider';
 
 export const DO_COLOR_DEFAULT_VALUE = { red: 0, green: 200, blue: 165, alpha: 1 };
 export const DONT_COLOR_DEFAULT_VALUE = { red: 255, green: 55, blue: 90, alpha: 1 };
 
+export const DosDontsBlockWrapper = ({ appBridge }: BlockProps) => {
+    const [blockSettings] = useBlockSettings<Settings>(appBridge);
+
+    const shouldFetchAssets =
+        blockSettings.mode === BlockMode.TEXT_AND_IMAGE ||
+        blockSettings.hasCustomDoIcon ||
+        blockSettings.hasCustomDontIcon;
+
+    const Provider = shouldFetchAssets ? AssetsProvider : NoAssetsProvider;
+
+    return (
+        <Provider appBridge={appBridge}>
+            <DosDontsBlock appBridge={appBridge} />
+        </Provider>
+    );
+};
+
 export const DosDontsBlock: FC<BlockProps> = ({ appBridge }) => {
     const [blockSettings, setBlockSettings] = useBlockSettings<Settings>(appBridge);
-    const { blockAssets, addAssetIdsToKey, deleteAssetIdsFromKey } = useBlockAssets(appBridge);
+    const { blockAssets, addAssetIdsToKey, deleteAssetIdsFromKey } = useContext(AssetsContext);
     const { openAssetChooser, closeAssetChooser } = useAssetChooser(appBridge);
     const isEditing = useEditorState(appBridge);
     const wrapperRef = useRef<HTMLDivElement>(null);
@@ -95,7 +112,7 @@ export const DosDontsBlock: FC<BlockProps> = ({ appBridge }) => {
     const columnGap = isCustomColumnGutter ? customColumnGutterValue : GUTTER_VALUES[columnGutterChoice];
     const rowGap = isCustomRowGutter ? customRowGutterValue : GUTTER_VALUES[rowGutterChoice];
     const sensors = useDndSensors(parseInt(columnGap ?? '0'), parseInt(rowGap ?? '0'));
-    const { dontIconAsset, doIconAsset, itemImages } = blockAssets;
+    const { dontIconAsset, doIconAsset, itemImages } = blockAssets ?? {};
     const [localItems, setLocalItems] = useState<Item[]>(items);
 
     const themeStyle = getComputedStyle(document.body);
@@ -189,19 +206,21 @@ export const DosDontsBlock: FC<BlockProps> = ({ appBridge }) => {
             });
             assetIds.push(image.id);
         }
-        addAssetIdsToKey('itemImages', assetIds).then(() => {
-            setAndSaveItems([...localItems, ...newItems]);
-            setIsUploadLoading(false);
-            setSelectedType(undefined);
-            setSelectedAssets(undefined);
-            setSelectedFiles(null);
-        });
+        if (addAssetIdsToKey) {
+            addAssetIdsToKey('itemImages', assetIds).then(() => {
+                setAndSaveItems([...localItems, ...newItems]);
+                setIsUploadLoading(false);
+                setSelectedType(undefined);
+                setSelectedAssets(undefined);
+                setSelectedFiles(null);
+            });
+        }
     };
 
     const removeItemById = (itemId: string) => {
         const itemToRemove = localItems.find((item) => item.id === itemId);
         const newItems: Item[] = localItems.filter((item) => item.id !== itemId);
-        if (itemToRemove?.imageId) {
+        if (itemToRemove?.imageId && deleteAssetIdsFromKey) {
             deleteAssetIdsFromKey('itemImages', [itemToRemove?.imageId]);
         }
         setAndSaveItems(newItems);
